requirements.txt
boto3==1.40.11
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-huggingface==0.3.1
langchain-mcp-adapters==0.1.9
langgraph==0.6.5
pydantic==2.11.7
PyYAML==6.0.2
streamlit==1.48.1
